revise CraftState scoring
Rationale:

Since our selection policy now evaluates nodes by average score *and*
max score, we likely no longer need an "all or nothing" scoring
strategy (either the craft reaches 100% progress with some amount
of quality, or it receives a score of zero).

The new scoring method provides a value even when the craft "fails",
which is OK since the max score should be enough to guide the search.
In testing, this revised scoring seems to give comparable results,
while also converging to a decent solution slightly sooner.

The fewer steps bonus has been removed, since it -- like the "all or
nothing" scoring -- only applied when the craft succeeded, which
means it didn't provide any signal for a bad path. Now, the progress
and quality bonuses should be competing with the durability and CP
bonuses, since a gain in progress or quality costs durability and CP.
This interaction should be enough to produce an efficient solution
organically.